Add new entities
Solve small bugs

Add automatic calculations refresh

+++ src/modules/vibi/vibi-core/src/main/java/it/geosolutions/opensdi2/persistence/MetricsDao.java
@@ -1,4 +1,4 @@
 package it.geosolutions.opensdi2.persistence;
 
-public interface MetricsDao extends GenericVibiDao<Metrics, Long> {
+public interface MetricsDao extends GenericVibiDao<Metrics, Integer> {
 }
